add generated code with yardl 0.6.4

+++ matlab/toolbox/+mrd/+binary/ImageHeaderSerializer.m
@@ -21,7 +21,7 @@ classdef ImageHeaderSerializer < yardl.binary.RecordSerializer
       field_serializers{16} = yardl.binary.OptionalSerializer(yardl.binary.Uint32Serializer);
       field_serializers{17} = yardl.binary.OptionalSerializer(yardl.binary.Uint64Serializer);
       field_serializers{18} = yardl.binary.VectorSerializer(yardl.binary.Uint64Serializer);
-      field_serializers{19} = yardl.binary.EnumSerializer('mrd.ImageType', @mrd.ImageType, yardl.binary.Uint64Serializer);
+      field_serializers{19} = yardl.binary.EnumSerializer('mrd.ImageType', @mrd.ImageType, yardl.binary.Int32Serializer);
       field_serializers{20} = yardl.binary.OptionalSerializer(yardl.binary.Uint32Serializer);
       field_serializers{21} = yardl.binary.OptionalSerializer(yardl.binary.Uint32Serializer);
       field_serializers{22} = yardl.binary.VectorSerializer(yardl.binary.Int32Serializer);
